test(private-cli): satisfy introduced clone gate
