+++ tests/private-hosted-cli/orchestrator.test.js
@@ -3,157 +3,10 @@
 const assert = require('node:assert/strict');
 const { afterEach, describe, it } = require('node:test');
 const {
-  HostedRunOrchestrator,
   RemoteAllocationUncertainError,
   RemoteDetachedError,
 } = require('../../private/hosted-cli-candidate/orchestrator');
-
-const RUNTIME_DIGEST = `sha256:${'a'.repeat(64)}`;
-const GRAPH = {
-  profile: 'openengine.graph.single-worker/v1',
-  root: { kind: 'step', worker: 'legacy.zeroshot.ship@1', attempts: 1 },
-};
-const CALLER_INPUT = Object.freeze({
-  source: 'prompt',
-  prompt: 'Ship the requested change.',
-  artifacts: [],
-});
-
-function base(overrides = {}) {
-  const sequence = [];
-  const requests = { plan: [], apply: [] };
-  let ids = 0;
-  const adapter = {
-    allocate() {
-      sequence.push('allocate');
-      return { id: 'cap1', state: 'ready' };
-    },
-    inspect() {
-      sequence.push('inspect');
-      return { id: 'cap1', state: 'ready' };
-    },
-    terminate() {
-      sequence.push('terminate');
-      return { id: 'cap1', state: 'terminating' };
-    },
-    ...overrides.adapter,
-  };
-  const initialClient = {
-    plan(params) {
-      sequence.push('plan');
-      requests.plan.push(params);
-      return { ok: true, diagnostics: [] };
-    },
-    apply(params) {
-      sequence.push('apply');
-      requests.apply.push(params);
-      return { generation: 1, runId: 'server-run-1', phase: 'running', deduped: false };
-    },
-    ...overrides.initialClient,
-  };
-  const finalClient = {
-    get() {
-      sequence.push('get');
-      return {
-        status: {
-          phase: 'finished',
-          observedGeneration: 1,
-          currentRunId: 'server-run-1',
-          atCursor: 'cursor-2',
-        },
-      };
-    },
-    ...overrides.finalClient,
-  };
-  let opens = 0;
-  const coordinator = {
-    open() {
-      sequence.push('initialize');
-      opens += 1;
-      return {
-        initializeResult: {
-          capabilities: { graphProfiles: ['openengine.graph.single-worker/v1'] },
-        },
-        client: opens === 1 ? initialClient : finalClient,
-      };
-    },
-    watch() {
-      sequence.push('watch');
-      let delivered = false;
-      return {
-        [Symbol.asyncIterator]() {
-          return this;
-        },
-        next() {
-          if (delivered) return { done: true };
-          delivered = true;
-          return {
-            done: false,
-            value: {
-              type: 'event',
-              runId: 'server-run-1',
-              cursor: 'cursor-2',
-              event: {
-                type: 'finished',
-                final_status: {
-                  phase: 'finished',
-                  observedGeneration: 1,
-                  currentRunId: 'server-run-1',
-                  atCursor: 'cursor-2',
-                },
-              },
-            },
-          };
-        },
-        cancel() {
-          sequence.push('watch-cancel');
-        },
-      };
-    },
-    close() {
-      sequence.push('close');
-    },
-    ...overrides.coordinator,
-  };
-  const output = { stdout: [], stderr: [] };
-  const orchestrator = new HostedRunOrchestrator({
-    assertGraphSpec: () => undefined,
-    readInputs: () => {
-      sequence.push('read-inputs');
-      return { graph: GRAPH, input: overrides.input ?? CALLER_INPUT };
-    },
-    checkHostedSetup: () => {
-      sequence.push('check-setup');
-      return { repository: 'owner/repo', provider: 'codex', modelLevel: 'level2' };
-    },
-    createCoordinator: () => coordinator,
-    randomUUID: () => `${String(++ids).padStart(8, '0')}-0000-0000-0000-000000000000`,
-    runtimeImageDigest: RUNTIME_DIGEST,
-    sleep: () => undefined,
-    output: {
-      stdout: (line) => output.stdout.push(line),
-      stderr: (line) => output.stderr.push(line),
-    },
-  });
-  return {
-    adapter,
-    coordinator,
-    orchestrator,
-    output,
-    requests,
-    sequence,
-    options: {
-      adapter,
-      target: { id: 'target1', url: 'https://target.example', organization: { id: 'org1' } },
-      graphPath: 'graph.json',
-      inputPath: 'input.json',
-      expectedRepository: 'owner/repo',
-      expectedProvider: 'codex',
-      expectedModelLevel: 'level2',
-      detach: false,
-    },
-  };
-}
+const { base, CALLER_INPUT } = require('./orchestrator-harness');
 
 it('refuses setup mismatches before allocation', async () => {
   const h = base();
