+++ tests/private-hosted-cli/services-lifecycle.test.js
@@ -2,348 +2,15 @@
 
 const assert = require('node:assert/strict');
 const { describe, it } = require('node:test');
-const { createDefaultServices } = require('../../private/hosted-cli-candidate/default-services');
 const { RemoteDetachedError } = require('../../private/hosted-cli-candidate/orchestrator');
-const { createTargetServices } = require('../../private/hosted-cli-candidate/target-services');
 const {
   HostedSessionCoordinator: RealHostedSessionCoordinator,
 } = require('../../lib/hosted-session/index.cjs');
 const { FakeWebSocket, settle } = require('../cluster/harness');
 const { makeAccess, waitForRequest, waitForSocket } = require('../hosted-session/harness');
-
-const RUNTIME_DIGEST = `sha256:${'a'.repeat(64)}`;
-const GRAPH = {
-  profile: 'openengine.graph.single-worker/v1',
-  root: { kind: 'step', worker: 'legacy.zeroshot.ship@1', attempts: 1 },
-};
-const DESCRIPTOR = {
-  origin: 'https://target.example',
-  oauth: {
-    deviceAuthorizationEndpoint: 'https://target.example/oauth/device',
-    tokenEndpoint: 'https://target.example/oauth/token',
-    revocationEndpoint: 'https://target.example/oauth/revoke',
-    clientId: 'private-candidate',
-    deviceGrantType: 'urn:ietf:params:oauth:grant-type:device_code',
-    audience: 'capsule',
-  },
-  capsule: { baseUrl: 'https://target.example/capsules/' },
-  session: { routeTemplate: { template: '/sessions/{capsuleId}' } },
-  sizes: { catalog: ['tiny', 'small', 'standard', 'large'], default: 'small' },
-};
-
-async function captureLogs(operation) {
-  const original = console.log;
-  const lines = [];
-  console.log = (...values) => lines.push(values.join(' '));
-  try {
-    const value = await operation();
-    return { lines, value };
-  } finally {
-    console.log = original;
-  }
-}
-
-function targetHarness() {
-  const calls = [];
-  const state = {
-    _targets: {
-      prod: {
-        id: 'target-prod',
-        url: 'https://target.example',
-        organization: { id: 'org-1' },
-        createdAt: '2026-08-03T00:00:00.000Z',
-      },
-    },
-  };
-  class TargetSessionManager {
-    constructor(options) {
-      calls.push(['session', options.targetName]);
-    }
-
-    login() {
-      calls.push(['login']);
-      return { organization: { id: 'org-1' } };
-    }
-
-    revoke(force) {
-      calls.push(['revoke', force]);
-    }
-  }
-  const runtime = {
-    target: {
-      TARGET_ACCOUNT: 'refresh-token',
-      TargetSessionManager,
-      acquireTargetLock: () => undefined,
-      addTarget(name, url, settings, descriptor) {
-        calls.push(['add', name, url, descriptor.origin]);
-        const record = {
-          id: `target-${name}`,
-          url,
-          createdAt: '2026-08-04T00:00:00.000Z',
-        };
-        settings.mutate((current) => {
-          current._targets[name] = record;
-        });
-        return record;
-      },
-      discoverTarget(url) {
-        calls.push(['discover', url]);
-        return DESCRIPTOR;
-      },
-      KeyringCredentialStore: {
-        create() {
-          return {
-            delete(service, account) {
-              calls.push(['delete', service, account]);
-            },
-          };
-        },
-      },
-      listTargets(settings) {
-        return Object.entries(settings.load()._targets).map(([name, record]) => ({ name, record }));
-      },
-      normalizeAndValidateUrl: (url) => url,
-      removeTarget(name, settings) {
-        calls.push(['remove', name]);
-        settings.mutate((current) => {
-          delete current._targets[name];
-        });
-      },
-      targetServiceKey: (id) => `zeroshot-target-${id}`,
-    },
-  };
-  const settings = {
-    load: () => state,
-    mutate: (mutator) => mutator(state),
-  };
-  const services = createTargetServices({
-    runtime,
-    settings,
-    httpTransport: () => ({ fetch: () => undefined }),
-    requireTarget: (name) => state._targets[name],
-  });
-  return { calls, services, state };
-}
-
-function remoteHarness(options = {}) {
-  const calls = [];
-  let ids = 0;
-  const target = {
-    id: 'target-prod',
-    url: 'https://target.example',
-    organization: { id: 'org-1' },
-    hostedSetup: {
-      kind: 'zeroshot.private-hosted-setup/v1',
-      repository: 'owner/repository',
-      provider: 'codex',
-      modelLevel: 'level2',
-      configuredAt: '2026-08-03T00:00:00.000Z',
-    },
-  };
-  const state = { _targets: { prod: target } };
-  const adapter = {
-    access(capsuleId, signal) {
-      const access = options.access?.(capsuleId, signal);
-      if (!access) throw new Error('unexpected capsule access request');
-      calls.push(['access', access.accessToken]);
-      return access;
-    },
-    allocate(params) {
-      calls.push(['allocate', params]);
-      if (options.allocationError) throw options.allocationError;
-      return {
-        id: 'cap-1',
-        state: 'ready',
-        label: params.label ?? null,
-        createdAt: '2026-08-04T00:00:00.000Z',
-      };
-    },
-    inspect(id) {
-      calls.push(['inspect', id]);
-      return { id, state: 'ready' };
-    },
-    list(params) {
-      calls.push(['list', params]);
-      return {
-        capsules: [
-          {
-            id: 'cap-1',
-            state: 'ready',
-            label: null,
-            createdAt: '2026-08-04T00:00:00.000Z',
-          },
-        ],
-        nextCursor: null,
-      };
-    },
-    terminate(id) {
-      calls.push(['terminate', id]);
-      return { id, state: 'terminating' };
-    },
-  };
-  let runOpenCount = 0;
-  class HostedSessionCoordinator {
-    constructor(init) {
-      calls.push(['coordinator', init.capsuleId, init.targetAuthority]);
-    }
-
-    open() {
-      calls.push(['initialize']);
-      runOpenCount += 1;
-      if (options.run) {
-        return {
-          initializeResult: {
-            capabilities: { graphProfiles: ['openengine.graph.single-worker/v1'] },
-          },
-          client:
-            runOpenCount === 1
-              ? {
-                  plan(params) {
-                    calls.push(['plan', params]);
-                    return { ok: true, diagnostics: [] };
-                  },
-                  apply(params) {
-                    calls.push(['apply', params]);
-                    return { generation: 1, runId: 'run-1' };
-                  },
-                }
-              : {
-                  get() {
-                    calls.push(['get']);
-                    return {
-                      status: {
-                        phase: 'finished',
-                        observedGeneration: 1,
-                        currentRunId: 'run-1',
-                        atCursor: 'cursor-1',
-                      },
-                    };
-                  },
-                },
-        };
-      }
-      return {
-        client: {
-          get() {
-            calls.push(['get']);
-            return {
-              status: {
-                phase: 'finished',
-                observedGeneration: 3,
-                currentRunId: 'run-3',
-                atCursor: 'cursor-3',
-              },
-            };
-          },
-          stop(params) {
-            calls.push(['stop', params]);
-            return { effectiveMode: params.mode, runId: 'run-3' };
-          },
-        },
-      };
-    }
-
-    watch(params) {
-      calls.push(['watch', params]);
-      if (options.interruptWatch) {
-        process.emit('SIGINT');
-        return {
-          [Symbol.asyncIterator]() {
-            return this;
-          },
-          next() {
-            return Promise.reject(params.signal.reason);
-          },
-          cancel() {
-            calls.push(['watch-cancel']);
-          },
-        };
-      }
-      let delivered = false;
-      return {
-        [Symbol.asyncIterator]() {
-          return this;
-        },
-        next() {
-          if (delivered) return { done: true };
-          delivered = true;
-          return {
-            done: false,
-            value: {
-              type: 'event',
-              runId: 'run-1',
-              cursor: 'cursor-1',
-              event: {
-                type: 'finished',
-                final_status: {
-                  phase: 'finished',
-                  observedGeneration: 1,
-                  currentRunId: 'run-1',
-                  atCursor: 'cursor-1',
-                },
-              },
-            },
-          };
-        },
-        cancel() {
-          calls.push(['watch-cancel']);
-        },
-      };
-    }
-
-    close() {
-      calls.push(['close']);
-    }
-  }
-  class TargetSessionManager {
-    tokenProvider() {
-      return () => Promise.resolve('access-token');
-    }
-  }
-  const runtime = {
-    cluster: { assertGraphSpec: () => undefined },
-    hostedSession: { HostedSessionCoordinator },
-    hostedTarget: {
-      createTargetAdapter(init) {
-        calls.push(['adapter', init.organization.id]);
-        return adapter;
-      },
-    },
-    target: {
-      TargetSessionManager,
-      discoverTarget() {
-        calls.push(['discover']);
-        return options.descriptor ?? DESCRIPTOR;
-      },
-      getTarget: (name) => state._targets[name],
-      KeyringCredentialStore: { create: () => ({}) },
-    },
-  };
-  const services = createDefaultServices({
-    createCoordinator: options.createCoordinator,
-    orchestratorOutput: options.orchestratorOutput,
-    runtime,
-    loadSettings: () => state,
-    mutateSettings: (mutator) => mutator(state),
-    httpTransport: () => ({ fetch: () => undefined }),
-    randomUUID: () => `${String(++ids).padStart(8, '0')}-0000-0000-0000-000000000000`,
-    manifest: {
-      privateMarker: 'ZEROSHOT_PRIVATE_HOSTED_CLI_CANDIDATE_DO_NOT_PUBLISH',
-      repository: 'owner/repository',
-      provider: 'codex',
-      modelLevel: 'level2',
-      runtimeImageDigest: RUNTIME_DIGEST,
-    },
-    readHostedInputs: () => {
-      calls.push(['read-inputs']);
-      return {
-        graph: GRAPH,
-        input: { source: 'prompt', prompt: 'Ship the change.', artifacts: [] },
-      };
-    },
-  });
-  return { adapter, calls, services };
-}
+const { captureLogs, DESCRIPTOR } = require('./candidate-fixtures');
+const { remoteHarness } = require('./remote-service-harness');
+const { targetHarness } = require('./target-service-harness');
 
 describe('private target services', () => {
   it('runs add, login, list, setup, and remove through production service wiring', async () => {
